chore: extract useStyleSheet
